check canary bytes impl

+++ src/include/fl.h
@@ -5,7 +5,7 @@
 #include <page.h>
 
 #define get_address(base, offset) (void*)((char*)base + offset)
-#define get_bin_index(internal_size) (int)(internal_size / CHUNK_ALIGNMENT - 3)
+#define get_bin_index(internal_size) (uint8_t)(internal_size / CHUNK_ALIGNMENT - 3)
 #define get_bin_size(index) (size_t)((size_t)(index + 3) * CHUNK_ALIGNMENT)
 
 #define get_bin_alloc_status(metadata)   (bool)((uintptr_t)metadata & 1)
